- getEstimatedAltitude completely rewritten and now it gives very good results in GUI.

+++ trunc/MultiWii/MultiWii.ino
@@ -46,7 +46,7 @@ December  2011     V1.dev
 #define PIDITEMS 8
 
 static uint32_t currentTime = 0;
-static uint16_t previousTime = 0;
+static uint32_t previousTime = 0;
 static uint16_t cycleTime = 0;     // this is the number in micro second to achieve a full loop, it can differ a little and is taken into account in the PID loop
 static uint16_t calibratingA = 0;  // the calibration is done is the main loop. Calibrating decreases at each cycle down to 0, then we enter in a normal mode.
 static uint8_t  calibratingM = 0;
